Updating drupal a bit
I've updated to use a better feature cut of our awestruct endpoints, one
which doesn't use pathauto which is very time consuming.

I also combined some of the extension enabling and downloading.

It would be much better if we could get all of this into the docker
container at build time instead of runtime.