Taking out logging on mysql
Logging is an extremely performance killing operation in mysql. I'm
removing it to help increase perf a little bit when we push to drupal.

Right now the whole build, assuming you've already built the containers,
takes about 10 minutes on my box. That's building the site and pushing
it into drupal. I have a feeling there's something else going on inside
drupal which is costing us a lot of time as I don't recall it being this
slow before.