Adding a couple new extensions for drupal
I'm doing the drupal push during a transform now instead of an extension
which happens before the transform. This could easily become an
after_generation extension in the next version of awestruct.

I'm also testing to see if drupal is alive before proceeding forward in
doing the push, this way simply using 'control.rb -r' will do the whole
build for us.