Doing checks in the control script for liveliness
The control script now checks to see if containers are active and ready
for data before we move on to the other parts of the build. I need
someone to test this on OS X to make sure things are working well for
boot2docker or other ways of running the docker container.