Docker script and containers are working
Finally everything seems to be working!
I switched to head Awestruct, will switch to the released version as
soon as I get some bugs ironed out.

I had to change the drupal volume location for some reason, don't know
why it wasn't working, but whatever.

I'm no longer using rvm as that seemed to be causing some issues, so I
dropped it. If we do change ruby versions we'll need to update the
containers.

Sending the pages to drupal adds an extra two minutes to the build. I
don't like it for development, but maybe it isn't that big of a deal,
not sure yet.